backend/requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
opencv-python
pdf2image